exercise/RestExercise/serializers.py
from rest_framework import serializers
from .models import Category, Item


class CategorySerializer(serializers.ModelSerializer):

    class Meta:
        model = Category
        fields = ['id', 'title']


class ItemSerializer(serializers.ModelSerializer):

    class Meta:
        model = Item
        fields = ['id', 'category', 'title', 'description', 'price']
